+++ client/src/Components/PaymentPage/Address.tsx
@@ -5,21 +5,10 @@ import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 import RadiusButton from 'Components/Common/RadiusButton';
 import { TYPE_getAddress } from 'Types/common/product';
 import TinyTitle from 'Components/Common/TinyTitle';
-import { ToastContainer, toast, Zoom } from 'react-toastify';
+import { toast } from 'react-toastify';
 import { useState, Dispatch, SetStateAction } from 'react';
 import 'react-toastify/dist/ReactToastify.css';
 import { useCustomMutation } from 'CustomHook/useCustomMutaiton';
-
-const StyleToastContainer = styled(ToastContainer)`
-  position: fixed;
-  top: 50%;
-  left: 50%;
-  -webkit-transform: translate(-50%, -50%);
-  -moz-transform: translate(-50%, -50%);
-  -ms-transform: translate(-50%, -50%);
-  -o-transform: translate(-50%, -50%);
-  transform: translate(-50%, -50%);
-`;
 
 const Title = styled.div`
   border-top: 1px solid black;
@@ -61,13 +50,11 @@ const Address: React.FC<Props> = ({
   );
   const sucessAlram = () =>
     toast.success('저장되었습니다.', {
-      position: 'top-right',
       hideProgressBar: false,
       closeOnClick: true,
       pauseOnHover: true,
       draggable: true,
       progress: undefined,
-      theme: 'light',
     });
 
   const onChangeForm = (e: React.ChangeEvent<HTMLInputElement>) => {
@@ -129,12 +116,6 @@ const Address: React.FC<Props> = ({
           setDataPut={setDataPut}
         />
       </div>
-      <StyleToastContainer
-        limit={4}
-        transition={Zoom}
-        hideProgressBar
-        autoClose={1000}
-      />
 
       <div className="my-5 relative h-6">
         <div className="absolute top-0 left-0">
