+++ client/src/Pages/PaymentPage.tsx
@@ -119,3 +119,4 @@ const PaymentPage: React.FC<{ session: any }> = ({ session }) => {
 };
 export default PaymentPage;
 //딜리트와 패치의 차이
+//패치를 할경우에 자동으로 겟이 다시됨 그렇지만 딜리트는 아니다
